+++ misc/client-extensions/ClientExtensions/Character/CharacterDefines.h
